+++ apps/cli/src/adapters/commander/index.ts
@@ -1,10 +1,12 @@
 import { Command } from "commander";
 
+import { Config } from "../../config.js";
 import { Dependencies } from "../../domain/dependencies.js";
 import { makeDoctorCommand } from "./commands/doctor.js";
+import { makeInfoCommand } from "./commands/info.js";
 import { makeVersionCommand } from "./commands/version.js";
 
-export const makeCli = (dependencies: Dependencies) => {
+export const makeCli = (deps: Dependencies, config: Config) => {
   const program = new Command();
 
   program
@@ -12,8 +14,9 @@ export const makeCli = (dependencies: Dependencies) => {
     .description("The CLI for DX-Platform")
     .version(__CLI_VERSION__);
 
-  program.addCommand(makeDoctorCommand(dependencies));
+  program.addCommand(makeDoctorCommand(deps, config));
   program.addCommand(makeVersionCommand());
+  program.addCommand(makeInfoCommand(deps, config));
 
   return program;
 };