+++ apps/cli/src/config.ts
@@ -1,0 +1,17 @@
+export type Config = {
+  minVersions: {
+    turbo: string;
+  };
+  repository: {
+    root: string;
+  };
+};
+
+export const getConfig = (repositoryRoot: string): Config => ({
+  minVersions: {
+    turbo: "2",
+  },
+  repository: {
+    root: repositoryRoot,
+  },
+});