+++ apps/cli/src/domain/validation.ts
@@ -4,17 +4,17 @@ export type ValidationCheck = FailedCheck | SuccessfulCheck;
 
 export type ValidationCheckResult = Result<ValidationCheck, Error>;
 
-export interface ValidationReporter {
-  reportValidationResult(result: ValidationCheckResult): void;
-}
+export type ValidationReporter = {
+  reportCheckResult(result: ValidationCheck): void;
+};
 
-interface FailedCheck extends Pick<SuccessfulCheck, "checkName"> {
+type FailedCheck = Pick<SuccessfulCheck, "checkName"> & {
   errorMessage: string;
   isValid: false;
-}
+};
 
-interface SuccessfulCheck {
+type SuccessfulCheck = {
   checkName: string;
   isValid: true;
   successMessage: string;
-}
+};