+++ infra/modules/github_selfhosted_runner_on_codebuild/_tests/setup/variables.tf
@@ -2,7 +2,7 @@ variable "environment" {
   type = object({
     prefix          = string
     env_short       = string
-    location        = string
+    region          = string
     domain          = optional(string)
     app_name        = string
     instance_number = string