+++ packages/azure-tracing/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @pagopa/azure-tracing
 
+## 0.4.6
+
+### Patch Changes
+
+- 25f7786: Replace `@tsconfig/node20` with `@tsconfig/node22`
+
+## 0.4.5
+
+### Patch Changes
+
+- ce98aa8: Bump dependencies to fix a peerDependency warning
+
 ## 0.4.4
 
 ### Patch Changes